restructure folder for tests
